CSIC-343 add read-only support to iscsi protocol only

+++ e2e/csitestimage/csitestimage.go
@@ -2,9 +2,11 @@ package main
 
 import (
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/signal"
+	"strconv"
 	"syscall"
 	"time"
 )
@@ -14,7 +16,22 @@ func main() {
 	fmt.Printf("current user id %d\n", os.Getuid())
 	fmt.Printf("current group id %d\n", os.Getgid())
 
-	fi, err := openFile()
+	readOnlyEnvVar := os.Getenv("READ_ONLY")
+	var readOnly bool
+	var err error
+
+	if readOnlyEnvVar != "" {
+		readOnly, err = strconv.ParseBool(readOnlyEnvVar)
+		if err != nil {
+			fmt.Printf("error parsing READ_ONLY env var %s", err.Error())
+			os.Exit(2)
+		}
+	}
+	if readOnly {
+		fmt.Println("READ_ONLY is true")
+	}
+
+	fi, err := openFile(readOnly)
 	if err != nil {
 		fmt.Printf("error opening file %s", err.Error())
 		os.Exit(2)
@@ -25,8 +42,12 @@ func main() {
 	fmt.Println("starting csitest...")
 
 	for {
-		writeToFile(fi)
-		time.Sleep(time.Second * 10)
+		if readOnly {
+			readFile(fi)
+		} else {
+			writeToFile(fi)
+		}
+		time.Sleep(time.Second * 30)
 	}
 
 }
@@ -45,21 +66,28 @@ func catchSignal() {
 
 }
 
-func openFile() (*os.File, error) {
+func openFile(readOnly bool) (*os.File, error) {
 	var fi *os.File
 	var fileName = "/tmp/csitesting/testfile"
 
 	_, err := os.Stat(fileName)
 	if err != nil {
-		fmt.Println("file does not exist, will create...")
-		fi, err = os.Create(fileName)
-		if err != nil {
-			fmt.Printf("error creating file %s\n", err.Error())
-			return nil, err
+		if !readOnly {
+			fmt.Println("file does not exist, will create...")
+			fi, err = os.Create(fileName)
+			if err != nil {
+				fmt.Printf("error creating file %s\n", err.Error())
+				return nil, err
+			}
 		}
 	} else {
 		fmt.Println("file already exists...")
-		fi, err = os.OpenFile(fileName, os.O_APPEND|os.O_WRONLY, os.ModeAppend)
+		if readOnly {
+			fi, err = os.Open(fileName)
+			fmt.Println("opening in read-only mode")
+		} else {
+			fi, err = os.OpenFile(fileName, os.O_APPEND|os.O_WRONLY, os.ModeAppend)
+		}
 		if err != nil {
 			fmt.Printf("error opening file %s\n", err.Error())
 			return nil, err
@@ -75,4 +103,24 @@ func writeToFile(fi *os.File) {
 		os.Exit(2)
 	}
 
+}
+
+func readFile(fi *os.File) {
+	var totalBytes int
+	buf := make([]byte, 1024)
+	for {
+		n, err := fi.Read(buf)
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			//fmt.Println(err)
+			continue
+		}
+		if n > 0 {
+			//fmt.Println(string(buf[:n]))
+			totalBytes = totalBytes + n
+		}
+	}
+	fmt.Printf("%d read\n", totalBytes)
 }
